Updated code code to account for locks, not very efficient, need to work on copy method

+++ filesystem.java
@@ -1,9 +1,10 @@
 import java.util.ArrayList;
 import java.util.Date;
+import java.util.concurrent.TimeUnit;
 
 /*
  * Operating Systems Lab Final
- * @author: Jen Zhu
+ * @author: Sadi Evren Seker, Jen Zhu
  * Due Date: 12/15/16
  *
  */
@@ -12,11 +13,20 @@ import java.util.Date;
 class filesystem {
 
     ArrayList<myFile> fs;
+    //ArrayList<sharedFile> currentlyAccessedFiles;
+    ArrayList<myFile> currentlyReading;
+    ArrayList<myFile> currentlyWriting;
+
 
     public filesystem() {
         this.fs = new ArrayList<myFile>();
+        //this.currentFiles = new ArrayList<sharedFile>();
+        this.currentlyReading = new ArrayList<myFile>();
+        this.currentlyWriting = new ArrayList<myFile>();
 
     } 
+
+    /** Helper Method to see if currentlyReading contains the file*/
 
     /** Method will create a new file */
     public boolean createFile(String name,int size, Date date){
@@ -30,11 +40,18 @@ class filesystem {
         return true;
     }
 
-    /** Method will write to a file */
+    /** Method will WRITE to a file */
+    // Note: Write will overwrite all information in the file, insert will add to existing
+    // content in the file, both will use writeLocks
+    // Open a file for writing while another is reading  -> NO
     public boolean write(String fileName, String content){
         for(myFile f : fs){
             if(f.name.equals(fileName)){
-                // call method from myFile here
+                if (currentlyReading.contains(f) || currentlyWriting.contains(f)){
+                    System.out.println("Unable to access file for writing.");
+                    return false;
+                }
+                currentlyWriting.add(f);
                 Date updatedDate = new Date();
                 try {
                     f.write(content, updatedDate);
@@ -48,11 +65,20 @@ class filesystem {
         return false;
     }
 
-    /** Method will read a file */
+    /** Method will READ a file */
+    // Open a file for reading while already reading -> OK
+    // Open a file for reading while another thread is writing -> NO
     public String read(String fileName){
         for(myFile f : fs){
             if(f.name.equals(fileName)){
                 // call method from myFile here
+                if (currentlyWriting.contains(f)){
+                    System.out.println("Unable to access file for reading.");
+                    return "Null. \nAn Error occurred, please try closing the file and trying again.";
+                }
+                if (!currentlyReading.contains(f)){
+                    currentlyReading.add(f);  
+                }
                 try {
                     String content = f.read();
                     return content;
@@ -65,7 +91,7 @@ class filesystem {
         return null;
     }
 
-    /** Method will list all the files in the directory */
+    /** Method will LIST all the files in the directory */
     public String list(){
         String returnval = "";
         for(myFile f : fs){
@@ -74,7 +100,11 @@ class filesystem {
         return returnval;
     }
 
-    /** Method will copy file contents into another file */
+    /** Method will COPY file contents into another file */
+    // Copy from file while another thread is writing to the file -> NO
+    // Copy from file while another is reading -> OK
+    // Copy into a file while another is reading -> NO
+    // Copy into file while another thread is writing to the file -> NO
     public boolean copy(String originalFileName, String fileCopiedToName){
         myFile originalFile = null;
         myFile fileCopiedTo = null;
@@ -85,18 +115,34 @@ class filesystem {
                 fileCopiedTo = f;
             }
         }
-        // question: are these instancees of the original files, or actually dealing with those objects?
+        if (currentlyReading.contains(fileCopiedTo) || currentlyWriting.contains(fileCopiedTo)){
+            System.out.println("Unbale to copy to file.");
+            return false;
+        } if (currentlyWriting.contains(originalFile)){
+            System.out.println("unable to copy from file.");
+            return false;
+        }
+        if (!currentlyReading.contains(originalFile)){
+            currentlyReading.add(originalFile);
+        } 
+        currentlyWriting.add(fileCopiedTo);
         fileCopiedTo.content = originalFile.content;
         return true;
     }
 
     /** Method will delete a file */
+    // Delete a file while another thread is reading/writing -> NO
+    // Delete a file while another thread is writing -> NO
     public void delete(String fileToDeleteName){
         myFile fileToDelete = null;
         for (myFile f:fs){
             if (f.name.equals(fileToDeleteName)){
                 fileToDelete = f;
             }
+        } 
+        if (currentlyWriting.contains(fileToDelete) || currentlyReading.contains(fileToDelete)){
+            System.out.println("Unable to delete. Please close all locks and try again.");
+            return;
         }
         fs.remove(fs.indexOf(fileToDelete));        
     }
@@ -105,6 +151,11 @@ class filesystem {
     public boolean insert(String fileToUpdate, String content){
         for (myFile f:fs){
             if (f.name.equals(fileToUpdate)){
+                if (currentlyReading.contains(f) || currentlyWriting.contains(f)){
+                    System.out.println("Unable to insert into the file.");
+                    return false;
+                }
+                currentlyWriting.add(f);
                 Date newDate = new Date();
                 try {
                     f.insert(content, newDate);
@@ -117,12 +168,52 @@ class filesystem {
         System.out.println("file not found");
         return false;
     }
+
+    /** Method to rename a file */
+    public boolean rename(String fileToRename, String newName) {
+        for (myFile f: fs){
+            if (f.name.equals(fileToRename)){
+                if (currentlyReading.contains(f) || currentlyWriting.contains(f)){
+                    return false;
+                }
+                f.rename(newName);
+                return true;
+            }
+        }
+        System.out.println("Unable to rename file.");
+        return false;
+    }
+
+    /** Method to Close locks on file */
+    public void close(String fileToClose){
+        //System.out.println("Entered the close method!");
+        for (myFile f: fs){
+            if (f.name.equals(fileToClose)){
+                // System.out.println("Found the file!");
+                f.close();
+                if (currentlyWriting.contains(f)) {
+                    currentlyWriting.remove(f);
+                } if (currentlyReading.contains(f)) {
+                    currentlyWriting.remove(f);
+                }
+            }
+        }
+    }
     /** Additional methods to be Addressed: */
     // locking in general (? how to test?)
-    // Move a file
-    // --> Need to find a way to store both arrayLists of files and files in filesystem
-    // --> assume both files exist, 
-    // Insert into a file (? adding to content in the file?)
-    // Close a file - close and remove all locks (? releasing the locks?)
+    //  --> create a data structure, synchronize via data structure
+    //  --> array keeping track of locks on numerous files and check permissions from array before opening file
+
+    // Locking Situations: 
+    // Open a file for reading while already reading -> OK
+    // Open a file for writing while another is reading  -> NO
+    // Copy a file while another is reading -> OK
+    // Delete a file while another thread is reading/writing -> NO
+    // Insert into a file while another thread is reading -> NO
+    // Open a file for reading while another thread is writing -> NO
+    // Delete a file while another thread is writing -> NO
+    // Copy from file while another thread is writing to the file -> NO
+    // Copy into file while another thread is writing to the file -> NO
+    // 
 
 }
